Tests + decent names for extracted methods/classes
